+++ src.mk
@@ -25,6 +25,7 @@ LIB_SOURCES =                                                   \
   cloud/cloud_scheduler.cc                                      \
   cloud/cloud_storage_provider.cc                               \
   cloud/cloud_file_cache.cc                                     \
+  cloud/cloud_file_deletion_scheduler.cc                        \
   db/arena_wrapped_db_iter.cc                                   \
   db/blob/blob_fetcher.cc                                       \
   db/blob/blob_file_addition.cc                                 \
